Hashtags and mentions in Russian tweets
The tweets dataframe containing the Russian tweets is already loaded. 
Compute the number of hashtags and mentions in each tweet by defining two functions count_hashtags() and count_mentions() respectively and applying them to the content feature of tweets. 
In case you don't recall, the tweets are contained in the content feature of tweets.